Ajout du calcul du Crc et de la méthode avancer dans la classe Robot. Et test d'ajout d'une nouvelle classe pour gérer l'appui des touches mais ça ne marche pas
