+++ Robot.h
@@ -1,0 +1,39 @@
+#ifndef MYROBOT_H
+#define MYROBOT_H
+
+#include <QObject>
+#include <QTcpSocket>
+#include <QAbstractSocket>
+#include <QDebug>
+#include <QTimer>
+#include <QMutex>
+#include <QWebEngineView>
+
+class MyRobot : public QObject {
+    Q_OBJECT
+public:
+    explicit MyRobot(QObject *parent = 0);
+    void doConnect();
+    void disConnect();
+    QByteArray DataToSend;
+    QByteArray DataReceived;
+    QMutex Mutex;
+    void test();
+    void crcToSend();
+
+signals:
+    void updateUI(const QByteArray Data);
+public slots:
+    void connected();
+    void disconnected();
+    void bytesWritten(qint64 bytes);
+    void readyRead();
+    void MyTimerSlot();
+    void avancer();
+
+private:
+    QTcpSocket *socket;
+    QTimer *TimerEnvoi;
+};
+
+#endif // MYROBOT_H
