+++ WifiBot2.pro
@@ -11,14 +11,16 @@ CONFIG += c++17
 #DEFINES += QT_DISABLE_DEPRECATED_BEFORE=0x060000    # disables all the APIs deprecated before Qt 6.0.0
 
 SOURCES += \
+    Robot.cpp \
     main.cpp \
     mainwindow.cpp
 
 HEADERS += \
+    Robot.h \
     mainwindow.h
 
 FORMS += \
-    mainwindow.ui
+    Robot.ui
 
 # Default rules for deployment.
 qnx: target.path = /tmp/$${TARGET}/bin
