+++ mainwindow.cpp
@@ -1,70 +1,12 @@
-// myrobot.cpp
-
 #include "mainwindow.h"
-
-MyRobot::MyRobot(QObject *parent) : QObject(parent) {
-    DataToSend.resize(9);
-    DataToSend[0] = 0xFF;
-    DataToSend[1] = 0x07;
-    DataToSend[2] = 120;
-    DataToSend[3] = 0x0;
-    DataToSend[4] = 120;
-    DataToSend[5] = 0x0;
-    DataToSend[6] = 80;
-    DataToSend[7] = 0x0;
-    DataToSend[8] = 0x0;
-    DataReceived.resize(21);
-    TimerEnvoi = new QTimer();
-    // setup signal and slot
-    connect(TimerEnvoi, SIGNAL(timeout()), this, SLOT(MyTimerSlot())); //Send data to wifibot timer
-}
+#include <windows.h>
+#include "Robot.h"
 
 
-void MyRobot::doConnect() {
-    socket = new QTcpSocket(this); // socket creation
-    connect(socket, SIGNAL(connected()),this, SLOT(connected()));
-    connect(socket, SIGNAL(disconnected()),this, SLOT(disconnected()));
-    connect(socket, SIGNAL(bytesWritten(qint64)),this, SLOT(bytesWritten(qint64)));
-    connect(socket, SIGNAL(readyRead()),this, SLOT(readyRead()));
-    qDebug() << "connecting..."; // this is not blocking call
-    //socket->connectToHost("LOCALHOST", 15020);
-    socket->connectToHost("192.168.10.1", 5001); // connection to wifibot
-    // we need to wait...
-    if(!socket->waitForConnected(5000)) {
-        qDebug() << "Error: " << socket->errorString();
-        return;
+void mainwindow::keyPressEvent(QKeyEvent *event)
+{
+    if (event->key() == Qt::Key_Z)
+    {
+        Robot.avancer();
     }
-    TimerEnvoi->start(75);
-
-}
-
-void MyRobot::disConnect() {
-    TimerEnvoi->stop();
-    socket->close();
-}
-
-void MyRobot::connected() {
-    qDebug() << "connected..."; // Hey server, tell me about you.
-}
-
-void MyRobot::disconnected() {
-    qDebug() << "disconnected...";
-}
-
-void MyRobot::bytesWritten(qint64 bytes) {
-    qDebug() << bytes << " bytes written...";
-}
-
-void MyRobot::readyRead() {
-    qDebug() << "reading..."; // read the data from the socket
-    DataReceived = socket->readAll();
-    emit updateUI(DataReceived);
-    qDebug() << DataReceived[0] << DataReceived[1] << DataReceived[2];
-}
-
-void MyRobot::MyTimerSlot() {
-    qDebug() << "Timer...";
-    while(Mutex.tryLock());
-    socket->write(DataToSend);
-    Mutex.unlock();
 }
